Refactors a bit of the program for a deeper understanding.
